transcode_jpeg: performance optimizations
- Replace unordered_map with flat array scratch buffer in
  ComputeSignallingOverhead, avoiding per-call allocation; split into
  per-cluster ComputeClusterSignallingOverhead for incremental updates
- Eliminate hist_h_a copy in RollbackScratch: recover pre-merge state
  by subtracting hist_h_b from the merged result instead of saving a
  full heap-allocated copy
- Incrementally track per-cluster header costs in the overhead-aware
  tail merge loop instead of recomputing the full overhead each iteration
- Eliminate div/mod by runtime M_eff in FlushTerm hot loop: replace
  with n_ci_start counter advancing by M_eff steps
- Short-circuit refine_iters==0 early inside RefineClustered to avoid
  redundant branching in callers; unify the call site in enc_jpeg_frame.cc
- Minor: remove PrepareIdentityBuckets, inline into PrepareEqualPopBuckets

+++ lib/jxl/transcode_jpeg/enc_jpeg_opt_data.cc
@@ -52,7 +52,7 @@ uint32_t JPEGOptData::CompactHBin(uint32_t zdc_ai) const {
 // The interval-cost matrix is written in the same diff form consumed by the
 // shared Knuth solver, so this stage reuses the exact same partition backend
 // as the AC-driven optimization path.
-Thresholds InitThresh(const JPEGOptData& d, int axis,
+Thresholds InitThresh(const JPEGOptData& d, uint32_t axis,
                       uint32_t target_intervals) {
   if (target_intervals <= 1) return {};
   const Thresholds& dc_vals = d.DC_vals[axis];
@@ -69,7 +69,7 @@ Thresholds InitThresh(const JPEGOptData& d, int axis,
   std::vector<uint32_t> prefix(M + 1, 0);
   for (uint32_t i = 0; i < M; ++i) {
     prefix[i + 1] =
-        prefix[i] + d.DC_cnt[axis][static_cast<uint32_t>(dc_vals[i] + kDCTOff)];
+        prefix[i] + d.DC_cnt[axis][dc_vals[i] + kDCTOff];
   }
 
   KnuthPartitionSolver solver(M);
